Updated SECURITY.md and removed imagine.py
