+++ SECURITY.md
@@ -2,20 +2,17 @@
 
 ## Supported Versions
 
-Use this section to tell people about which versions of your project are
-currently being supported with security updates.
+The most tested version of the code is going to be the most recent x.0 release. Any other release is going to be
+considered still in development, though may still be supported depending on the level of testing.
 
 | Version | Supported          |
 | ------- | ------------------ |
-| 5.1.x   | :white_check_mark: |
-| 5.0.x   | :x:                |
-| 4.0.x   | :white_check_mark: |
-| < 4.0   | :x:                |
+| 1.0     | :white_check_mark: |
+| 1.1     | :x:                |
 
 ## Reporting a Vulnerability
 
-Use this section to tell people how to report a vulnerability.
-
-Tell them where to go, how often they can expect to get an update on a
-reported vulnerability, what to expect if the vulnerability is accepted or
-declined, etc.
+Create an issue using the [GitHub issue feature](https://github.com/RIT-MDRC/MDRCBot/issues) and add the current 
+maintainer (found in the file's main docstring) of the file you found an issue with to that issue. Response times can
+very depending on the time of year, but if you do not see a response with a month, feel free to add any of the other
+authors to the issue to and they may be able to contact the maintainer directly.
